+++ community-cookbooks/maven/resources/default.rb
@@ -3,7 +3,7 @@
 # Resource:: default
 #
 # Author:: Bryan W. Berry <bryan.berry@gmail.com>
-# Copyright 2012, Opscode Inc.
+# Copyright 2012, Chef Software Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.