+++ community-cookbooks/maven/CHANGELOG.md
@@ -2,20 +2,32 @@ maven Cookbook CHANGELOG
 ========================
 This file is used to list changes made in each version of the maven cookbook.
 
+v1.3.0
+------
+- Adding Windows support
+
+v1.2.0
+------
+- Adding flag to allow Java not to be managed by cookbook
+
+v1.1.0
+------
+[COOK-3849] - Update maven 3 to 3.1.1
+
 
 v1.0.0
 ------
 ### Improvement
-- **[COOK-3470](https://tickets.opscode.com/browse/COOK-3470)** - Improve `/etc/mavenrc` template
-- **[COOK-3459](https://tickets.opscode.com/browse/COOK-3459)** - Install Maven 3.1.0 by default
+- **[COOK-3470](https://tickets.chef.io/browse/COOK-3470)** - Improve `/etc/mavenrc` template
+- **[COOK-3459](https://tickets.chef.io/browse/COOK-3459)** - Install Maven 3.1.0 by default
 
 v0.16.4
 -------
 ### Improvement
-- **[COOK-3352](https://tickets.opscode.com/browse/COOK-3352)** - Improve `repository_root` attribute customization
+- **[COOK-3352](https://tickets.chef.io/browse/COOK-3352)** - Improve `repository_root` attribute customization
 
 ### Bug
-- **[COOK-2799](https://tickets.opscode.com/browse/COOK-2799)** - Fix idempotency in LWRP
+- **[COOK-2799](https://tickets.chef.io/browse/COOK-2799)** - Fix idempotency in LWRP
 
 v0.16.2
 -------